Add temporary patch for linux installations.
It was discovered that when attempting to install `cb` via Homebrew on
Linux, the build process was failing. This failure was due to the
development dependency on `ameba`. This dependency has since been
removed in the currently unreleased/development version. So, to ensure
that `cb` can be built on Linux via Homebrew we are temporarily applying
the same patch to the v3.3.3 version as is applied in the upstream
development version.