Update formula to use prebuilt binaries.
Now that GH actions provides an Apple Silicon runner we can now simply point to
the artifacts that are generated as part of the release process on the official
GH repo for `cb`. So here we update the formula to no longer build the bottles
and instead pull the binary from the official releases.