v3.6.4 - remove bottle
It was reported in issue #32 that upgrading to v3.6.4 was failing due to
the bottle being missing. This is due to us having moved away from
packaging bottles in the traditional brew sense and instead pulling the
prebuilt binaries[1] that are a result of the official release process.

It's unclear why the bottle directive wasn't removed on this most recent
release however we're correcting that here.

[1] https://github.com/CrunchyData/bridge-cli/releases